Do not test MSHRs with prefetch fills
The occupancy of MSHRs is already tested with
test/cpp/src/402-miss-latency.cc and there is nothing special with
prefetch fills. This means we can safely remove the tests of MSHRs with
prefetch fills to save the code to maintain without hurting the ability
of the test suite to catch bugs.